Updates to address test lint errors
